+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
+- Improved logging when a user with orders is attempted to be deleted. ([#3686](https://github.com/craftcms/commerce/issues/3686))
+- Fixed a PHP error that could occur when calculating tax adjustments. ([#3822](https://github.com/craftcms/commerce/issues/3822))
+- Fixed a PHP error that could occur when updating an order’s status on console requests. ([#3858](https://github.com/craftcms/commerce/issues/3858))
+- Fixed a bug where additional tax ID validators where not being checked when a store’s "Validate Business Tax ID as Vat ID" setting was on.
+- Fixed a bug where the deprecated `TaxRate::$isVat` property was still being set. ([#3874](https://github.com/craftcms/commerce/issues/3874))
+- Fixed a bug where emails could be rendered in the incorrect language when sent manually from the control panel. ([#3852](https://github.com/craftcms/commerce/issues/3852))
+- Fixed an error that occurred when rendering a Link field with a product selected on the front end. ([#3833](https://github.com/craftcms/commerce/issues/3833))
+
 ## 5.3.0.2 - 2025-01-31
 
 - Fixed a bug where gateways weren’t ensuring handle uniqueness. ([#3839](https://github.com/craftcms/commerce/issues/3839))